moved experience logic to ExperienceManager
